add solution num3

+++ app/main.py
@@ -1,4 +1,4 @@
-def calculate_team_total_rating(players: list) -> float:
+def calculate_team_total_rating(players: list) -> int:
     return sum(player.get_rating() for player in players)
